[MAYBE] magit-completion-read: add 2TH-DEF argument
This argument is only used if DEF is nil and bypasses the new
mechanism that allows user to skip having to make a choice or
at least confirm the default.  It is useful when it is safe
to use the primary default without a change to abort, but not
to do so for some secondary fallback default.

This will be used by `magit-read-stash', which already does
something similar.  It might also be useful to treat the current
branch as such a secondary default.